requirements.txt
google-api-python-client
requests
boto3
google-cloud-bigquery